requirements.txt
pypika==0.48.9
Flask==3.0.0
yoyo-migrations==8.2.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9